chore(playground): hygiene — shared noop, stale label, agents page split
- ActivityFixtures.tsx / GoalBarFixtures.tsx: drop local NOOP redefinitions,
  reuse the shared noop already exported from PlaygroundComposerActions.tsx.
- ChatPlaygroundPage.tsx: fix stale tab-label leftover ("Typography
  hierarchy review" -> "Chat playground").
- AgentsPlaygroundPage.tsx: split into a thin page delegating to
  components/playground/AgentsSettingsPlayground.tsx, matching every other
  playground route's page/component split.
